Make model modular, ease model override
